rename SNG -> Logo
