+++ src/renderer/icons/Logo.vue
