+++ src/renderer/views/Library.vue
@@ -13,11 +13,11 @@
     <div
       class="w-full h-[72px] bg-[#FCF1DF] flex items-center justify-center p-6"
     >
-      <SNG class="inline-block h-6 w-6" />
+      <Logo class="inline-block h-6 w-6" />
       <span class="ml-4 text-xl">Slovenská národná galéria</span>
     </div>
   </div>
 </template>
 <script setup lang="ts">
-import SNG from "../icons/SNG.vue";
+import Logo from "../icons/Logo.vue";
 </script>
